openai_agent_witguardrails/pyproject.toml
[project]
name = "openai-agent-witguardrails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.19",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
]
